Refactoring, deleted file
